// src/openapi-types.ts
export declare namespace OpenAPIV3 {
  interface Document {
    openapi: string;
    info: InfoObject;
    servers?: ServerObject[];
    paths: PathsObject;
    components?: ComponentsObject;
    security?: SecurityRequirementObject[];
    tags?: TagObject[];
    externalDocs?: ExternalDocumentationObject;
  }
  interface InfoObject {
    title: string;
    description?: string;
    termsOfService?: string;
    contact?: ContactObject;
    license?: LicenseObject;
    version: string;
  }
  interface ContactObject {
    name?: string;
    url?: string;
    email?: string;
  }
  interface LicenseObject {
    name: string;
    url?: string;
  }
  interface ServerObject {
    url: string;
    description?: string;
    variables?: {
      [variable: string]: ServerVariableObject;
    };
  }
  interface ServerVariableObject {
    enum?: string[];
    default: string;
    description?: string;
  }
  interface PathsObject {
    [pattern: string]: PathItemObject;
  }
// ...
  interface ExternalDocumentationObject {
    description?: string;
    url: string;
  }
// ...
  type NonArraySchemaObjectType =
    | "null"
    | "boolean"
    | "object"
    | "number"
    | "string"
    | "integer";
  type ArraySchemaObjectType = "array";
  type SchemaObject = ArraySchemaObject | NonArraySchemaObject;
  interface ArraySchemaObject extends BaseSchemaObject {
    type: ArraySchemaObjectType;
    items: SchemaObject;
  }
  interface NonArraySchemaObject extends BaseSchemaObject {
    type: NonArraySchemaObjectType;
  }
  interface BaseSchemaObject {
    title?: string;
    description?: string;
    format?: string;
    default?: any;
    multipleOf?: number;
    maximum?: number;
    exclusiveMaximum?: boolean;
    minimum?: number;
    exclusiveMinimum?: boolean;
    maxLength?: number;
    minLength?: number;
    pattern?: string;
    additionalProperties?: boolean | SchemaObject;
    maxItems?: number;
    minItems?: number;
    uniqueItems?: boolean;
    maxProperties?: number;
    minProperties?: number;
    required?: string[];
    enum?: any[];
    properties?: {
      [name: string]: SchemaObject;
    };
    allOf?: Array<SchemaObject>;
    oneOf?: Array<SchemaObject>;
    anyOf?: Array<SchemaObject>;
    not?: SchemaObject;
    nullable?: boolean;
    discriminator?: DiscriminatorObject;
    readOnly?: boolean;
    writeOnly?: boolean;
    xml?: XMLObject;
    externalDocs?: ExternalDocumentationObject;
    example?: any;
    deprecated?: boolean;
  }
// ...
  interface SecurityRequirementObject {
    [name: string]: string[];
  }
  interface ComponentsObject {
    schemas?: {
      [key: string]: SchemaObject;
    };
    responses?: {
      [key: string]: ResponseObject;
    };
    parameters?: {
      [key: string]: ParameterObject;
    };
    examples?: {
      [key: string]: ExampleObject;
    };
    requestBodies?: {
      [key: string]: RequestBodyObject;
    };
    headers?: {
      [key: string]: HeaderObject;
    };
    securitySchemes?: {
      [key: string]: SecuritySchemeObject;
    };
    links?: {
      [key: string]: LinkObject;
    };
    callbacks?: {
      [key: string]: CallbackObject;
    };
  }
// ...
  interface TagObject {
    name: string;
    description?: string;
    externalDocs?: ExternalDocumentationObject;
  }
}